Move typedefs out of utils and into typing file.
