+++ src/ogd/common/interfaces/outerfaces/DictionaryOuterface.py
@@ -7,7 +7,7 @@ from ogd.common.interfaces.outerfaces.DataOuterface import DataOuterface
 from ogd.common.models.enums.ExportMode import ExportMode
 from ogd.common.schemas.configs.GameSourceSchema import GameSourceSchema
 from ogd.common.utils.Logger import Logger
-from ogd.common.utils.utils import ExportRow
+from ogd.common.utils.typing import ExportRow
 
 class DictionaryOuterface(DataOuterface):
