+++ src/ogd/common/utils/Readme.py
@@ -13,7 +13,7 @@ from typing import Any, Dict, IO, List, Optional, Set
 from ogd.common.schemas.games.GameSchema import GameSchema
 from ogd.common.schemas.tables.TableSchema import TableSchema
 from ogd.common.utils.Logger import Logger
-from ogd.common.utils.utils import ExportRow
+from ogd.common.utils.typing import ExportRow
 
 class Readme:
     def __init__(self, game_schema:GameSchema, table_schema:TableSchema):
