+++ src/ogd/common/utils/typing.py
@@ -1,3 +1,4 @@
 from typing import Any, Callable, Dict, List, Optional, TypeAlias
 
-Map : TypeAlias = Dict[str, Any] # type alias: we'll call any dict using string keys a "Map"
+Map       : TypeAlias = Dict[str, Any] # type alias: we'll call any dict using string keys a "Map"
+ExportRow : TypeAlias = List[Any]
